split out individual components of dashboard in order to make mocking work for tests
